chore: replaced all audio bridge plugin examples with a single example
