+++ jarust_plugins/examples/audio_bridge.rs
@@ -2,8 +2,7 @@ use jarust::jaconfig::JaConfig;
 use jarust::jaconfig::TransportType;
 use jarust::TransactionGenerationStrategy;
 use jarust_plugins::audio_bridge::jahandle_ext::AudioBridge;
-use jarust_plugins::audio_bridge::msg_opitons::CreateRoomOptions;
-use jarust_plugins::audio_bridge::msg_opitons::JoinRoomOptions;
+use jarust_plugins::audio_bridge::msg_opitons::MuteOptions;
 use std::path::Path;
 use tracing_subscriber::EnvFilter;
 
@@ -16,10 +15,11 @@ async fn main() -> anyhow::Result<()> {
     tracing_subscriber::fmt().with_env_filter(env_filter).init();
 
     let timeout = std::time::Duration::from_secs(10);
-    let capacity = 32;
     let config = JaConfig::builder()
         .url("ws://localhost:8188/ws")
-        .capacity(capacity)
+        .capacity(
+            32, /* Buffer size on the entire connection with janus */
+        )
         .build();
     let mut connection = jarust::connect(
         config,
@@ -28,40 +28,23 @@ async fn main() -> anyhow::Result<()> {
     )
     .await?;
     let session = connection.create(10, timeout).await?;
-    let (handle, mut event_receiver) = session.attach_audio_bridge(capacity).await?;
-
-    let create_room_rsp = handle
-        .create_room_with_config(
-            CreateRoomOptions {
-                secret: Some("superdupersecret".to_string()),
-                ..Default::default()
-            },
-            timeout,
-        )
+    let (handle, mut events) = session
+        .attach_audio_bridge(16 /* Buffer size on this handle only */)
         .await?;
 
-    tracing::info!(
-        "Created Room {:#?}, permanent: {}",
-        create_room_rsp.room,
-        create_room_rsp.permanent
-    );
+    let create_room_rsp = handle.create_room(None, timeout).await?;
+    let rooms = handle.list_rooms(timeout).await?;
+
+    tracing::info!("Rooms {:#?}", rooms);
 
     handle
         .join_room(
             create_room_rsp.room.clone(),
-            JoinRoomOptions {
-                secret: Some("superdupersecret".to_string()),
-                generate_offer: Some(true),
-                ..Default::default()
-            },
+            Default::default(),
             None,
             timeout,
         )
         .await?;
-
-    if let Some(event) = event_receiver.recv().await {
-        tracing::info!("Joined Room {:#?}, {:#?}", create_room_rsp.room, event);
-    }
 
     let list_participants_rsp = handle
         .list_participants(create_room_rsp.room, timeout)
@@ -72,10 +55,28 @@ async fn main() -> anyhow::Result<()> {
         list_participants_rsp.participants
     );
 
-    handle.leave(timeout).await?;
+    use jarust_plugins::audio_bridge::events::AudioBridgeEvent as ABE;
+    use jarust_plugins::audio_bridge::events::PluginEvent as PE;
+    if let Some(PE::AudioBridgeEvent(ABE::RoomJoined { id, room, .. })) = events.recv().await {
+        handle
+            .mute(MuteOptions {
+                id: id.clone(),
+                room: room.clone(),
+                secret: None,
+            })
+            .await?;
 
-    if let Some(event) = event_receiver.recv().await {
-        tracing::info!("Event {:#?}", event);
+        handle
+            .unmute(MuteOptions {
+                id,
+                room,
+                secret: None,
+            })
+            .await?;
+    };
+
+    while let Some(e) = events.recv().await {
+        tracing::info!("{e:#?}");
     }
 
     Ok(())
